Initial work on fixing many gosec and linting issues

+++ internal/gateway/auth/cli_only_test.go
@@ -233,7 +233,7 @@ func TestCLIOnlyMiddlewareWithAPIToken(t *testing.T) {
 	expiresAt := time.Now().Add(24 * time.Hour)
 	tokenReq := &token.APITokenRequest{
 		Name:        "Test Token",
-		UserID:      int64(testUser.ID),
+		UserID:      testUser.ID,
 		Permissions: []string{"*"},
 		ExpiresAt:   &expiresAt,
 	}
